+++ src/main/kotlin/com/github/leomillon/uuidgenerator/settings/uuid/form/UUIDGeneratorSettingsForm.kt
@@ -1,29 +1,45 @@
 package com.github.leomillon.uuidgenerator.settings.uuid.form
 
 import com.github.leomillon.uuidgenerator.UUIDGenerator
+import com.github.leomillon.uuidgenerator.settings.UUIDGeneratorBaseForm
 import com.github.leomillon.uuidgenerator.settings.uuid.UUIDGeneratorSettings
-import java.util.UUID
+import java.awt.ItemSelectable
+import java.util.*
 import javax.swing.JCheckBox
 import javax.swing.JComponent
 import javax.swing.JLabel
 import javax.swing.JPanel
 import javax.swing.JRadioButton
+import javax.swing.JSpinner
 
-class UUIDGeneratorSettingsForm {
+class UUIDGeneratorSettingsForm : UUIDGeneratorBaseForm {
+
+    override var panel: JPanel? = null
+
+    override var uuidV4RadioButton: JRadioButton? = null
+    override var uuidV7RadioButton: JRadioButton? = null
+    override var lowerCaseRadioButton: JRadioButton? = null
+    override var upperCaseRadioButton: JRadioButton? = null
+    override var withDashesRadioButton: JRadioButton? = null
+    override var withoutDashesRadioButton: JRadioButton? = null
+    override var longSizeRadioButton: JRadioButton? = null
+    override var shortSizeRadioButton: JRadioButton? = null
+
+    override var currentTimeRadioButton: JRadioButton? = null
+    override var fixedTimeRadioButton: JRadioButton? = null
+    override var timePanel: JPanel? = null
+    override var yearSpinner: JSpinner? = null
+    override var monthSpinner: JSpinner? = null
+    override var daySpinner: JSpinner? = null
+    override var hourSpinner: JSpinner? = null
+    override var minuteSpinner: JSpinner? = null
+    override var secondSpinner: JSpinner? = null
+    override var millisSpinner: JSpinner? = null
 
     private val uuidV4Preview = UUID.fromString("242e6506-710e-4af9-8b3d-dfcfad4bc9a6")
     private val uuidV7Preview = UUID.fromString("019d10e9-3e28-7683-9758-8d7a97205099")
 
-    private var panel: JPanel? = null
     private var previewValue: JLabel? = null
-    private var uuidVersion4RadioButton: JRadioButton? = null
-    private var uuidVersion7RadioButton: JRadioButton? = null
-    private var lowerCaseRadioButton: JRadioButton? = null
-    private var upperCaseRadioButton: JRadioButton? = null
-    private var withDashesRadioButton: JRadioButton? = null
-    private var withoutDashesRadioButton: JRadioButton? = null
-    private var longSizeRadioButton: JRadioButton? = null
-    private var shortSizeRadioButton: JRadioButton? = null
     private var highlightingCheckbox: JCheckBox? = null
 
     private val settings: UUIDGeneratorSettings = UUIDGeneratorSettings.instance
@@ -33,8 +49,8 @@ class UUIDGeneratorSettingsForm {
     }
 
     fun loadSettings() {
-        uuidVersion4RadioButton?.isSelected = settings.uuidVersion4
-        uuidVersion7RadioButton?.isSelected = !settings.uuidVersion4
+        uuidV4RadioButton?.isSelected = settings.uuidVersion4
+        uuidV7RadioButton?.isSelected = !settings.uuidVersion4
         lowerCaseRadioButton?.isSelected = settings.lowerCased
         upperCaseRadioButton?.isSelected = !settings.lowerCased
         withDashesRadioButton?.isSelected = settings.withDashes
@@ -42,19 +58,40 @@ class UUIDGeneratorSettingsForm {
         longSizeRadioButton?.isSelected = settings.longSize
         shortSizeRadioButton?.isSelected = !settings.longSize
         highlightingCheckbox?.isSelected = settings.codeHighlighting
+        currentTimeRadioButton?.isSelected = !settings.fixedTime
+        fixedTimeRadioButton?.isSelected = settings.fixedTime
 
-        sequenceOf(
-            uuidVersion4RadioButton,
-            uuidVersion7RadioButton,
+        setPanelEnabled(timePanel, settings.fixedTime && !settings.uuidVersion4)
+
+        yearSpinner?.value = settings.year
+        monthSpinner?.value = settings.month
+        daySpinner?.value = settings.day
+        hourSpinner?.value = settings.hour
+        minuteSpinner?.value = settings.minute
+        secondSpinner?.value = settings.second
+        millisSpinner?.value = settings.millis
+
+        sequenceOf<ItemSelectable?>(
+            uuidV4RadioButton,
+            uuidV7RadioButton,
             lowerCaseRadioButton,
             upperCaseRadioButton,
             withDashesRadioButton,
             withoutDashesRadioButton,
             longSizeRadioButton,
-            shortSizeRadioButton
+            shortSizeRadioButton,
+            currentTimeRadioButton,
+            fixedTimeRadioButton
         )
             .filterNotNull()
-            .forEach { it.addItemListener { updatePreview() } }
+            .forEach { uiComponent: ItemSelectable? ->
+                uiComponent?.addItemListener {
+                    if (isUuid7TimeComponent(uiComponent)) {
+                        setPanelEnabled(timePanel, isFixedTime() && isUuidVersion7())
+                    }
+                    updatePreview()
+                }
+            }
 
         updatePreview()
     }
@@ -76,20 +113,35 @@ class UUIDGeneratorSettingsForm {
         settings.withDashes = isWithDashes() ?: true
         settings.longSize = isLongSize() ?: true
         settings.codeHighlighting = codeHighlightingEnabled() ?: true
+        settings.fixedTime = isFixedTime()
+
+        settings.year = year()
+        settings.month = month()
+        settings.day = dayOfMonth()
+        settings.hour = hour()
+        settings.minute = minute()
+        settings.second = second()
+        settings.millis = millis()
     }
 
     fun component(): JComponent? = panel
 
-    private fun isUuidVersion4() = uuidVersion4RadioButton?.isSelected
-    private fun isLowerCased() = lowerCaseRadioButton?.isSelected
-    private fun isWithDashes() = withDashesRadioButton?.isSelected
-    private fun isLongSize() = longSizeRadioButton?.isSelected
     private fun codeHighlightingEnabled() = highlightingCheckbox?.isSelected
 
     val isModified: Boolean
         get() = (isUuidVersion4() != settings.uuidVersion4
-            || isLowerCased() != settings.lowerCased
-            || isWithDashes() != settings.withDashes
-            || isLongSize() != settings.longSize
-            || codeHighlightingEnabled() != settings.codeHighlighting)
+                || isLowerCased() != settings.lowerCased
+                || isWithDashes() != settings.withDashes
+                || isLongSize() != settings.longSize
+                || codeHighlightingEnabled() != settings.codeHighlighting
+                || isFixedTime() != settings.fixedTime
+
+                || year() != settings.year
+                || month() != settings.month
+                || dayOfMonth() != settings.day
+                || hour() != settings.hour
+                || minute() != settings.minute
+                || second() != settings.second
+                || millis() != settings.millis
+                )
 }